Add files via upload
This is a Tableau project. The problem statement for the project is: An app based cab company finds that there is huge cancellation rate for a particular week. The objective of this project is to analysis the data and find out the factors for high cancellation and the operation area where its high. With the help of Tableau, I analyse the data and presented my findings in a client friendly dashboard and story presentation for easy understanding.